+++ src/slogo/model/backEndInternal/commands/PenDown.java
@@ -2,7 +2,7 @@ package slogo.model.backEndInternal.commands;
 
 import slogo.model.backEndInternal.BackEndTurtle;
 
-public class PenDown implements Command<Integer> {
+public class PenDown implements Command<Double> {
 
   BackEndTurtle backEndTurtle;
 
@@ -11,8 +11,8 @@ public class PenDown implements Command<Integer> {
   }
 
   @Override
-  public double execute() {
+  public Double execute() {
     backEndTurtle.flipPen();
-    return 1;
+    return 1.0;
   }
 }
